Refactor: split parsing and struct definitions
