+++ src/lib.rs
@@ -1,5 +1,7 @@
-mod bsii_decode;
+pub mod bsii_file;
 mod bsii_output;
-mod scsc_decrypt;
+mod bsii_parse;
+pub mod scsc_file;
+pub mod scsc_parse;
 #[cfg(feature = "wasm")]
 mod wasm;
