+++ src/main.rs
@@ -1,13 +1,15 @@
 use std::env;
 use std::fs;
 
-use bsii_decode::BsiiFile;
-use scsc_decrypt::ScscFile;
+use bsii_file::BsiiFile;
+use scsc_file::ScscFile;
 use simple_logger::SimpleLogger;
 
-mod bsii_decode;
+mod bsii_file;
 mod bsii_output;
-mod scsc_decrypt;
+mod bsii_parse;
+mod scsc_file;
+mod scsc_parse;
 
 fn main() {
     SimpleLogger::new()
@@ -17,9 +19,9 @@ fn main() {
 
     let arg = env::args().next_back().unwrap();
     let content = fs::read(arg).unwrap();
-    let scsc_file = ScscFile::from_content(content.as_slice()).unwrap();
-    let bsii_binary = scsc_file.to_bsii_binary();
+    let scsc_file = ScscFile::parse(content.as_slice()).unwrap();
+    let bsii_binary = scsc_file.decode();
     fs::write("output.bsii", &bsii_binary).unwrap();
-    let bsii_file = BsiiFile::from_content(&bsii_binary).unwrap();
-    println!("{}", bsii_output::bsii_to_siin(&bsii_file));
+    let bsii_file = BsiiFile::parse(&bsii_binary).unwrap();
+    println!("{}", bsii_file.to_siin());
 }
